interning symbols: modify symbol table to store MalDatum
Currently symbols are forever stored in the symbol table.
But we want to drop them from the table when they are no longer used.
To do that we need to count references, and we already have that in
MalDatum. To avoid adding ref count to Symbol, let's just use MalDatum.

Work in progress. Still unable to properly free symbols - a nasty memory
bug somewhere.